debugging

+++ Tower-Defense/Grid.h
@@ -4,7 +4,6 @@
 #include <fstream>
 #include <iostream>
 
-#include "Entity.h"
 #include "BlockArea.h"
 
 class Grid
@@ -17,7 +16,7 @@ private:
 public:
 	// store road
 	std::vector<std::pair<Position, Position>> & getRoad(int selectedRoad)  { return allRoads[selectedRoad]; };
-	int allRoadsSize() { return allRoads.size(); };
+	size_t allRoadsSize() { return allRoads.size(); };
 	
 	// load road
 	void saveRoad();
